Update README.md
One of Tableau's standout features is its unparalleled flexibility. For furniture showrooms, this means the ability to tailor dashboards to specific needs and preferences. Whether it's adjusting charts to focus on sales, quantity, or profit, or fine-tuning filters to explore regional, city, or state-level data, Tableau puts the power of customization in the hands of the user.